Add html form tag
Add input and label tags to form

Added inputs and labels for entering an username, email and password.
Also added a submit button at the end.

Add logic to create new user

Inside the create method, it checks if the user entered valid data.
If it is, it redirects to the 'new' view. Otherwise, it renders the
same view but with the data still on the fields.

Refactor create method

The create method now uses a new user_params that returns the values
needed to create a new user from a hash. This hash, called 'user', was
also created in the view and gets sent when the user submits the data.