Make a form tag with #form_with helper
Commented the previous form tag. Use again the User.new(user_params) in
the User Controller to use the 'user' hash. The #new now creates a new
User object and assigns it to the user instance variable. This is to use
it in the new form tag.

Using the #form_with, the form tag is recreated again, but this time
with a few extra options, like a different label value that's displayed
on the view and predefined values on the username and email input
fields. This was done to test the available options offered by these
helper methods to create the needed form.