Create view to edit users
Added routes to allow the #edit and #update actions needed for editing a
user's data.

Created and added code in #edit and #update actions to handle a user's
update.

Also created a new view to render the page when trying to update the
user. The code it uses is the same as the 'new' view. The only addition
is that when the user fails to meet the validations, it shows how many
errors are and list them with their explanations.